Revert "Set up to find cause"
This reverts commit 4343043b93da66ecb96fc82cff454710718ba535.
