+++ src/classic/clvm_tools/stages/stage_2/compile.rs
@@ -21,10 +21,7 @@ use crate::classic::clvm_tools::stages::stage_2::helpers::{brun, evaluate, quote
 use crate::classic::clvm_tools::stages::stage_2::module::compile_mod;
 use crate::compiler::srcloc::Srcloc;
 
-use crate::classic::clvm_tools::stages::stage_2::module::Indent;
-use crate::compiler::sexp::decode_string;
-
-const DIAG_OUTPUT: bool = false;
+const DIAG_OUTPUT: bool = true;
 
 lazy_static! {
     static ref PASS_THROUGH_OPERATORS: HashSet<Vec<u8>> = {
@@ -549,10 +546,8 @@ fn find_symbol_match<A: ClassicAllocator>(
 where
     A::NodePtr: Clone,
 {
-    let indent = Indent::new();
     if let Some(symlist) = proper_list(allocator, symbol_table, true) {
         for sym in symlist {
-            eprintln!("{indent}sym {}", allocator.disassemble(&sym, None));
             if let Some(symdef) = proper_list(allocator, &sym, true) {
                 if symdef.is_empty() {
                     continue;
@@ -666,7 +661,6 @@ fn compile_application<A: ClassicAllocator>(
 where
     A::NodePtr: Clone,
 {
-    let indent = Indent::new();
     let mut compiled_args = vec![operator.clone()];
 
     let loc = allocator.loc(prog);
@@ -717,9 +711,7 @@ where
             if PASS_THROUGH_OPERATORS.contains(opbuf) || (!opbuf.is_empty() && opbuf[0] == b'_') {
                 Ok(r)
             } else {
-                eprintln!("{indent}find symbol {} in\n{indent}{}\n{indent}{}", decode_string(opbuf), allocator.disassemble(&prog, None), allocator.disassemble(&symbol_table, None));
-                find_symbol_match(allocator, opbuf, &r, symbol_table).and_then(|x| {
-                match x {
+                find_symbol_match(allocator, opbuf, &r, symbol_table).and_then(|x| match x {
                     Some(SymbolResult::Direct(v)) => Ok(v),
                     Some(SymbolResult::Matched(_symbol, value)) => {
                         let loc = allocator.loc(&value);
@@ -752,7 +744,7 @@ where
                         enlist(allocator, &[apply_atom, value, cons_enlisted])
                     }
                     None => error_result,
-                }})
+                })
             }
         }
         None => error_result,
@@ -770,10 +762,9 @@ pub fn do_com_prog<A: ClassicAllocator>(
 where
     A::NodePtr: Clone,
 {
-    let indent = Indent::new();
     if DIAG_OUTPUT {
         println!(
-            "START COMPILE {}: {}\n{indent}MACRO {}\n{indent}SYMBOLS {}",
+            "START COMPILE {}: {}\nMACRO {}\nSYMBOLS {}",
             from,
             allocator.disassemble(prog, None),
             allocator.disassemble(macro_lookup, None),
@@ -783,7 +774,7 @@ where
     do_com_prog_(allocator, prog, macro_lookup, symbol_table, run_program).inspect(|x| {
         if DIAG_OUTPUT {
             println!(
-                "DO_COM_PROG {}: {}\n{indent}MACRO {}\n{indent}SYMBOLS {}\n{indent}RESULT {}",
+                "DO_COM_PROG {}: {}\nMACRO {}\nSYMBOLS {}\nRESULT {}",
                 from,
                 allocator.disassemble(prog, None),
                 allocator.disassemble(macro_lookup, None),
@@ -804,7 +795,6 @@ fn do_com_prog_<A: ClassicAllocator>(
 where
     A::NodePtr: Clone,
 {
-    let indent = Indent::new();
     /*
      * Turn the given program `prog` into a clvm program using
      * the macros to do transformation.
@@ -860,7 +850,6 @@ where
                                         symbol_table,
                                         run_program.clone()
                                     ).and_then(|x| x.map(Ok).unwrap_or_else(|| m! {
-                                        let _ = eprintln!("{indent}compile_application {} {}\n{indent}{}", allocator.disassemble(&operator, None), allocator.disassemble(&prog, None), allocator.disassemble(&prog_rest, None));
                                         compile_application(
                                             allocator,
                                             &prog,
